Move CertificateExpirationPolicy
