+++ api/src/main/java/io/strimzi/api/kafka/model/CertificateExpirationPolicy.java
@@ -2,7 +2,7 @@
  * Copyright 2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.api.kafka;
+package io.strimzi.api.kafka.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
